+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # HPO Workbench Changelog
 
 ## v0.1.9
+
+### hpowbcore
 * Support for Word export (discontinuing support for RTF export)
+
+### hpowbgui
 * Adding help dialog
+* Display HPO Annotations according to organ system
+* More detailed RTD documentation on excel, github issues, and browsing
 
 
 ## v0.1.8